+++ shoppingcart-api/src/sql/trigger/trSubject.trigger.sql
@@ -13,7 +13,7 @@ CREATE TRIGGER trSubject
 	SET NOCOUNT ON
 
 	DECLARE 
-		@AuditLogTransactionId	int,
+		@AuditLogTransactionId	bigint,
 		@Inserted	    		int = 0,
  		@ROWS_COUNT				int
